@@ -3,7 +3,7 @@ name = "SpectraFit"
 version = "1.4.0"
 description = "Fast fitting of 2D- and 3D-Spectra with established routines"
 authors = [{ name = "Anselm Hahn", email = "anselm.hahn@gmail.com" }]
-requires-python = ">=3.8,<3.14"
+requires-python = ">=3.9,<3.14"
 readme = "README.md"
 license = "BSD-3-Clause"
 maintainers = [{ name = "Anselm Hahn", email = "anselm.hahn@gmail.com" }]
@@ -160,8 +167,8 @@ build-backend = "hatchling.build"
 [tool.ruff]
 fix = true
 include = ["*.py"]
-src = ["spectrafit"]
-target-version = "py38"
+src = ["spectrafit", "test"]
+target-version = "py39"
 
 [tool.ruff.lint]
 select = ["ALL"]
