@@ -5,15 +5,12 @@ from __future__ import annotations
 import gzip
 import json
 import pickle
-import shutil
 import sys
 
 from pathlib import Path
 from typing import TYPE_CHECKING
 from typing import Any
-from typing import Dict
 from typing import Literal
-from typing import Tuple
 
 import numpy as np
 import pandas as pd
@@ -28,19 +25,14 @@ from spectrafit.plugins.file_converter import FileConverter
 from spectrafit.plugins.pkl_converter import ExportData
 from spectrafit.plugins.pkl_converter import PklConverter
 from spectrafit.plugins.pkl_visualizer import PklVisualizer
+from spectrafit.plugins.pptx_converter import PPTXConverter
+from spectrafit.plugins.rixs_converter import RIXSConverter
 
 
 if TYPE_CHECKING:
     from matplotlib.figure import Figure
     from numpy.typing import NDArray
     from pytest_console_scripts import ScriptRunner
-
-
-if sys.version_info >= (3, 9):
-    from spectrafit.plugins.pptx_converter import PPTXConverter
-    from spectrafit.plugins.rixs_converter import RIXSConverter
-else:
-    pytest.mark.skip("Requires Python 3.9 or higher", allow_module_level=True)
 
 
 @pytest.mark.skipif(sys.version_info < (3, 9), reason="Requires Python 3.9 or higher")
@@ -420,7 +412,7 @@ class TestDataConverter:
 
 
 @pytest.fixture(autouse=True, name="tmp_file_dict")
-def tmp_file_dict() -> Dict[str, Dict[str, NDArray[Any]]]:
+def tmp_file_dict() -> dict[str, dict[str, NDArray[Any]]]:
     """Create temporary file with pickle data."""
     return {
         "level_1": {
@@ -439,7 +431,7 @@ def tmp_file_dict() -> Dict[str, Dict[str, NDArray[Any]]]:
 
 
 @pytest.fixture(autouse=True, name="tmp_file_nested_dict")
-def tmp_file_nested_dict() -> Dict[str, Dict[str, Any]]:
+def tmp_file_nested_dict() -> dict[str, dict[str, Any]]:
     """Create temporary file with nested pickle data."""
     return {
         "level_4": {
@@ -462,7 +454,7 @@ def tmp_file_nested_dict() -> Dict[str, Dict[str, Any]]:
 @pytest.fixture(autouse=True, name="tmp_pkl_gz_file")
 def tmp_pkl_gz_file(
     tmp_path: Path,
-    tmp_file_dict: Dict[str, Dict[str, NDArray[np.float64]]],
+    tmp_file_dict: dict[str, dict[str, NDArray[np.float64]]],
 ) -> Path:
     """Create temporary file with pickle data.
 
@@ -483,7 +475,7 @@ def tmp_pkl_gz_file(
 @pytest.fixture(autouse=True, name="tmp_file_pkl")
 def tmp_file_pkl(
     tmp_path: Path,
-    tmp_file_dict: Dict[str, Dict[str, NDArray[np.float64]]],
+    tmp_file_dict: dict[str, dict[str, NDArray[np.float64]]],
 ) -> Path:
     """Create temporary file with pickle data.
 
@@ -504,7 +496,7 @@ def tmp_file_pkl(
 @pytest.fixture(autouse=True, name="tmp_file_pkl_gz")
 def tmp_file_pkl_gz(
     tmp_path: Path,
-    tmp_file_dict: Dict[str, Dict[str, NDArray[np.float64]]],
+    tmp_file_dict: dict[str, dict[str, NDArray[np.float64]]],
 ) -> Path:
     """Create temporary file with pickle data.
 
@@ -525,7 +517,7 @@ def tmp_file_pkl_gz(
 @pytest.fixture(autouse=True, name="tmp_file_pkl_nested")
 def tmp_file_pkl_nested(
     tmp_path: Path,
-    tmp_file_nested_dict: Dict[str, Dict[str, Any]],
+    tmp_file_nested_dict: dict[str, dict[str, Any]],
 ) -> Path:
     """Create temporary file with nested pickle data.
 
@@ -554,7 +546,7 @@ class TestPklConverter:
     def test_export_data(
         self,
         tmp_path: Path,
-        tmp_file_dict: Dict[str, Dict[str, NDArray[np.float64]]],
+        tmp_file_dict: dict[str, dict[str, NDArray[np.float64]]],
         export_format: str,
     ) -> None:
         """Test export data.
@@ -598,7 +590,7 @@ class TestPklConverter:
 
     def test_numpy2list(
         self,
-        tmp_file_dict: Dict[str, Dict[str, NDArray[np.float64]]],
+        tmp_file_dict: dict[str, dict[str, NDArray[np.float64]]],
     ) -> None:
         """Test numpy2list.
 
@@ -665,7 +657,7 @@ class TestPklConverter:
     def test_pkl_converter_nested(
         self,
         tmp_path: Path,
-        tmp_file_nested_dict: Dict[str, Any],
+        tmp_file_nested_dict: dict[str, Any],
     ) -> None:
         """Test pickle converter for nested dictionaries.
 
@@ -815,7 +807,7 @@ class TestPklAsGraph:
 @pytest.fixture(autouse=True, name="tmp_list_dict_rixs")
 def fixture_tmp_list_dict_rixs(
     tmp_path: Path,
-) -> Tuple[Path, Tuple[str, str, str]]:
+) -> tuple[Path, tuple[str, str, str]]:
     """Fixture for temporary list of dictionaries.
 
     Args:
@@ -850,7 +842,7 @@ class TestRixsConverter:
 
     def test_convertet(
         self,
-        tmp_list_dict_rixs: Tuple[Path, Tuple[str, str, str]],
+        tmp_list_dict_rixs: tuple[Path, tuple[str, str, str]],
     ) -> None:
         """Test the converter.
 
@@ -871,7 +863,7 @@ class TestRixsConverter:
     @pytest.mark.parametrize("export_format", ["json", "toml", "lock", "npy", "npz"])
     def test_save(
         self,
-        tmp_list_dict_rixs: Tuple[Path, Tuple[str, str, str]],
+        tmp_list_dict_rixs: tuple[Path, tuple[str, str, str]],
         export_format: str,
     ) -> None:
         """Test the save function for various export formats.
@@ -913,7 +905,7 @@ class TestRixsConverter:
 
     def test_raise_error_save(
         self,
-        tmp_list_dict_rixs: Tuple[Path, Tuple[str, str, str]],
+        tmp_list_dict_rixs: tuple[Path, tuple[str, str, str]],
     ) -> None:
         """Test the raise error.
 
@@ -932,7 +924,7 @@ class TestRixsConverter:
     @pytest.mark.parametrize("mode", ["sum", "mean"])
     def test_create_rixs(
         self,
-        tmp_list_dict_rixs: Tuple[Path, Tuple[str, str, str]],
+        tmp_list_dict_rixs: tuple[Path, tuple[str, str, str]],
         mode: str,
     ) -> None:
         """Test the create rixs.
@@ -956,7 +948,7 @@ class TestRixsConverter:
 
     def test_create_rixs_fail_1(
         self,
-        tmp_list_dict_rixs: Tuple[Path, Tuple[str, str, str]],
+        tmp_list_dict_rixs: tuple[Path, tuple[str, str, str]],
     ) -> None:
         """Test the create rixs fail.
 
@@ -983,8 +975,8 @@ class TestRixsConverter:
     )
     def test_create_rixs_fail_2(
         self,
-        tmp_list_dict_rixs: Tuple[Path, Tuple[str, str, str]],
-        keys: Tuple[str, str, str],
+        tmp_list_dict_rixs: tuple[Path, tuple[str, str, str]],
+        keys: tuple[str, str, str],
     ) -> None:
         """Test the create rixs fail.
 
@@ -1010,7 +1002,7 @@ class TestRixsConverter:
     def test_cmd_pkl_converter(
         self,
         script_runner: Any,
-        tmp_list_dict_rixs: Tuple[Path, Tuple[str, str, str]],
+        tmp_list_dict_rixs: tuple[Path, tuple[str, str, str]],
         export_format: str,
     ) -> None:
         """Test the data converter plugin.
@@ -1093,7 +1085,7 @@ gaussian_1 = [0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8]
 """
 
 
-@pytest.mark.skipif(sys.version_info < (3, 9), reason="Requires Python 3.9 or higher")
+# @pytest.mark.skipif(sys.version_info < (3, 9), reason="Requires Python 3.9 or higher")
 class TestPPTXConverter:
     """Test the pptx converter."""
 
@@ -1128,10 +1120,7 @@ class TestPPTXConverter:
             file_format (str): File format.
 
         """
-        source_path = Path("spectrafit/plugins/img/SpectraFit.png")
-        destination_path: Path = PPTXBasicTitleAPI().credit_logo
-        destination_path.parent.mkdir(parents=True, exist_ok=True)
-        shutil.copy(source_path, destination_path)
+        _ = PPTXBasicTitleAPI().credit_logo
 
         ret = script_runner.run(
             [
@@ -1141,7 +1130,6 @@ class TestPPTXConverter:
                 file_format,
             ],
         )
-        destination_path.unlink()
         assert ret.success
         assert ret.stderr == ""
         assert ret.stdout == ""
