@@ -6,7 +6,6 @@ import sys
 
 from typing import TYPE_CHECKING
 from typing import Any
-from typing import Tuple
 
 import numpy as np
 import plotly.graph_objects as go
@@ -19,18 +18,15 @@ if TYPE_CHECKING:
     from numpy.typing import NDArray
 
 
-if sys.version_info >= (3, 9):
-    from spectrafit.plugins.rixs_converter import RIXSConverter
-    from spectrafit.plugins.rixs_visualizer import RIXSApp
-    from spectrafit.plugins.rixs_visualizer import RIXSFigure
-    from spectrafit.plugins.rixs_visualizer import RIXSVisualizer
-else:
-    pytest.mark.skip("Requires Python 3.9 or higher", allow_module_level=True)
+from spectrafit.plugins.rixs_converter import RIXSConverter
+from spectrafit.plugins.rixs_visualizer import RIXSApp
+from spectrafit.plugins.rixs_visualizer import RIXSFigure
+from spectrafit.plugins.rixs_visualizer import RIXSVisualizer
 
 
 @pytest.mark.skipif(sys.version_info < (3, 9), reason="Requires Python 3.9 or higher")
 @pytest.fixture(scope="module", autouse=True, name="test_data")
-def fixture_test_data() -> Tuple[
+def fixture_test_data() -> tuple[
     NDArray[np.float64],
     NDArray[np.float64],
     NDArray[np.float64],
@@ -96,7 +92,7 @@ class TestRIXSApp:
         self,
         file_format: str,
         tmp_path: Path,
-        test_data: Tuple[NDArray[np.float64], NDArray[np.float64], NDArray[np.float64]],
+        test_data: tuple[NDArray[np.float64], NDArray[np.float64], NDArray[np.float64]],
     ) -> None:
         """Test the loading of data."""
         data = {
